Create README.md
Project developing in process. In the end I expect it to be a town building simulator (Kind of SimCity on PC, but stilistically closer to NimbleBit's BitCity), all the code, shaders and models have been done myself, excluding some parts of code that I've peeped on LearnOpenGL website of Joey de Vries (https://learnopengl.com/About). The font and textures were found over the Internet on different resources, although in future I would like to make my own. 
For now, resource management is hardcoded to using files on my own laptop's filesystem and the entire project is being developed with Linux on mind, so, if you want to have fun with it, make sure you have your copy of project configured properly on your PC and that you have all the dependent libs and frameworks setup(such as GLFW, GLEW, DevIL, FreeType, Assimp). Yes, it is uncomfortable, but again, my project is still in progress, all the OS dependent parts would be done in future.
Below are some cute screenshots I've made during the work, so you could see what it all looks like:
![2018-03-30 12-36-07](https://user-images.githubusercontent.com/29004250/38732683-bebb072e-3f27-11e8-821a-073351ade07c.png)
![2018-03-30 12-36-21](https://user-images.githubusercontent.com/29004250/38732691-c55391e6-3f27-11e8-9289-681cb9155b2b.png)
![2018-03-30 12-36-36](https://user-images.githubusercontent.com/29004250/38732695-ca1d910e-3f27-11e8-8463-8e415a317845.png)
![2018-03-30 12-37-47](https://user-images.githubusercontent.com/29004250/38732698-cc812aaa-3f27-11e8-8ef6-426119fe90a2.png)
![2018-04-06 18-07-13](https://user-images.githubusercontent.com/29004250/38732703-d18a1be2-3f27-11e8-9f83-1396bf0c4367.png)